Add key press to start game and game logic
The game starts when a key press is detected, showing the first color
of the sequence. The user must click the same color in order to progress
to the next level. The game then shows the next color and the user must
remember to click the previous sequence of colors before clicking the
latest one.